COMP: Update g3log external project to support reuse from application
This commit updates the project to specify the install commands
responsible to copy file to CMAKE_INSTALL_PREFIX only if *no*
custom install directory was specified.

It updates the logic to use g3log_INSTALL_DIR instead of CMAKE_INSTALL_PREFIX.

It ensures the install subdirectory exists when custom location is
specified.