[Backport MR-772] COMP: Update Assimp to simplify packaging in client projects
The install rules associated with the assimp target are now consistently
associated with the same component on macOS, Linux and Windows.

Prior to this commit, the component name used on macOS and Linux
was "Unspecified" instead of the value associated with the "LIBASSIMP_COMPONENT
CMake variable.

List of changes:

$ git shortlog ce3d6a91d..9c4296940 --no-merges
Jean-Christophe Fillion-Robin (1):
      [Backport PR-4391] cmake: Fix assimp target install rule fully specifying component